+++ ports/stm32/stm_jpeg.c
@@ -33,6 +33,7 @@
 #include STM32_HAL_H
 #include "irq.h"
 #include "stm_dma.h"
+#include "stm_mdma.h"
 
 #define JPEG_CODEC_TIMEOUT          (1000)
 #define JPEG_ALLOC_PADDING          ((__SCB_DCACHE_LINE_SIZE) * 4)
